Update auto_save.py
Prevent call to save if auto complete is visible. 

Note This change causes the save not to get called when auto complete popup is visible. So any file changes made while auto complete is visible is not saved.

This is a solution for the issue discussed in [https://github.com/jamesfzhang/auto-save/issues/21]  and [https://github.com/jamesfzhang/auto-save/issues/7]